pyproject.toml
[tool.poetry]
name = "turkey_rahyab_interview"
version = "0.1.0"
description = ""
authors = ["mehdi <mhd.mosavi@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "4.0"
djangorestframework = "^3.14.0"
Markdown = "^3.4.4"
django-filter = "^23.2"
django-rest-captcha = "^0.3.0"
Pillow = "^10.0.0"
django-imagekit = "^4.1.0"
djangorestframework-simplejwt = "^5.2.2"
ipython = "^8.14.0"
ipdb = "^0.13.13"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
